2long

+++ cc/932.c
@@ -2,7 +2,7 @@
 #include <stdio.h>
 int main()
 {
-    long T16, num;
+    long long T16, num;
     printf("Hello, World!\n");
     T16=0;
     for (num=11; num<1000000; num++)
